+++ migrations/migrate-properties.js
@@ -144,7 +144,7 @@ let migrateSnapshots = function() {
 }
 
 runMigrations = function() {
-    if(db.systemSnapshot.count({deltaFiProperties: {$exists: 1}}) == 0) {
+    if(db.propertySet.count() > 0 && db.systemSnapshot.count({deltaFiProperties: {$exists: 1}}) == 0) {
         migratePropertySet()
         migrateSnapshots()
     }